+++ src/components/button.js
@@ -1,0 +1,7 @@
+const createLoadMoreButton = function () {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
+
+export {createLoadMoreButton};
